webhook/requirements.txt
flask>=2.0.0
requests>=2.28.0
gspread>=5.0.0
google-auth>=2.0.0
gunicorn>=21.0.0
